improve code, fix errors
