+++ src/app/App.tsx
@@ -5,19 +5,13 @@ import { theme } from '@/shared/config/theme/theme.ts';
 import * as S from './App.styled.ts';
 import NotFound from '@/pages/NotFound.tsx';
 import Profile from '@/pages/Profile.tsx';
-import MainLayout from '@/layouts/MainLayout.tsx';
 import AuthLayout from '@/layouts/AuthLayout.tsx';
+import { MainLayout } from '@/layouts/MainLayout.tsx';
 import { PostList } from '@/widgets/PostList/ui/PostList.tsx';
-import { PostAddForm } from '@/widgets/PostAdd/ui/PostAddForm.tsx';
 import { Route, Routes } from 'react-router-dom';
 import { useEffect } from 'react';
 
 function App() {
-  const modalRef = useRef<HTMLDialogElement>(null);
-
-  const openPostAddForm = () => {
-    modalRef.current?.showModal();
-  };
   useEffect(() => {
     console.log(account);
   }, []);
@@ -29,9 +23,6 @@ function App() {
           {/* Общий лэйаут */}
           <Route element={<MainLayout />}>
             <Route path="/" element={<PostList />} />
-            <S.PostAddDialog ref={modalRef}>
-              <PostAddForm onClose={() => modalRef.current?.close()} />
-            </S.PostAddDialog>
           </Route>
 
           {/* Auth-лэйаут */}
